style: 코드를 기능별 분류 1. 이전에 있던 main.swift를 기능별로 분류하고 여러파일에 나눠 담음 2. 사용자 정의타입, 가위바위보, 묵찌빠, 공통기능 이라는 기준점을 사용함
