+++ RockPaperScissors/RockPaperScissors/GameResult.swift
@@ -1,0 +1,5 @@
+enum GameResult: String {
+    case win = "이겼습니다!"
+    case draw = "비겼습니다!"
+    case lose = "졌습니다!"
+}
