+++ RockPaperScissors/RockPaperScissors/Hand.swift
@@ -1,0 +1,5 @@
+enum Hand: Int {
+    case scissors = 1
+    case rock
+    case paper
+}
