+++ RockPaperScissors/RockPaperScissors/Turn.swift
@@ -1,0 +1,9 @@
+struct Turn {
+    var winner: String = ""
+    
+    mutating func inputWinner(winner: String) {
+        self.winner = winner
+    }
+}
+
+var turn = Turn()
